requirements.txt
fastapi
uvicorn[standard]
python-multipart
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv
opensearch-py>=2.4.0
langchain==0.3.27
langchain-core>=0.3.76
langchain-openai>=0.3.1
langchain-community>=0.3.15
langchain-mcp-adapters==0.1.11
nest_asyncio
tqdm
pydantic
mcp-server-opensearch
